Fix all actual tests at the moment except for User repository

+++ src/main/java/com/kirak/repository/datajpa/DataJpaAptTypeRepository.java
@@ -24,7 +24,7 @@ public interface DataJpaAptTypeRepository extends JpaRepository<AptType, Short> 
     @Modifying
     @Transactional
     @Query("DELETE FROM AptType t WHERE t.id=:id")
-    short delete(@Param("id") short id);
+    int delete(@Param("id") short id);
 
     @Override
     AptType findOne(Short id);
